@@ -1,4 +1,4 @@
-use std::{env, fs::File, path::Path, io::Write};
+use std::{env, fs::File, io::Write, path::Path};
 
 fn main() {
     let out_dir = env::var("OUT_DIR").unwrap();
