@@ -2,16 +2,17 @@
 
 use std::fmt::Display;
 
-use anyhow::Context;
-use http_body_util::Full;
-use hyper::body::Bytes;
+use anyhow::{Context, Error};
+use http_body_util::{combinators::BoxBody, BodyExt, Full, StreamBody};
+use hyper::body::{Bytes, Frame};
 use serde::{Deserialize, Serialize};
+use tokio::sync::mpsc::{unbounded_channel, UnboundedSender};
 
-use crate::{HttpResponse, APPLICATION_JSON, CONTENT_TYPE};
+use crate::{map_infallible, HttpResponse, Response, APPLICATION_JSON, CONTENT_TYPE};
 
 /// Universal API interface returns data format
 #[derive(Serialize, Deserialize, Debug)]
-// #[serde(rename_all = "camelCase")]
+#[serde(rename_all = "camelCase")]
 pub struct ApiResult<T> {
     /// result code (usually, 200 represents success, and 500 represents failure)
     pub code: u32,
@@ -28,7 +29,6 @@ pub struct Resp;
 
 impl<T> ApiResult<T> {
     /// Generate an ApiResult that represents success using the specified data
-    #[inline]
     pub fn ok(data: T) -> Self {
         Self {
             code: 200,
@@ -38,8 +38,7 @@ impl<T> ApiResult<T> {
     }
 
     /// Generate an ApiResult that represents success using empty data
-    #[inline]
-    pub fn ok_with_empty() -> Self {
+    pub fn ok_empty() -> Self {
         Self {
             code: 200,
             message: None,
@@ -48,7 +47,6 @@ impl<T> ApiResult<T> {
     }
 
     /// Generate an ApiResult indicating failure using the specified error message
-    #[inline]
     pub fn fail(msg: String) -> Self {
         Self {
             code: 500,
@@ -58,7 +56,6 @@ impl<T> ApiResult<T> {
     }
 
     /// Generate an ApiResult representing a failure using the specified error code and error message
-    #[inline]
     pub fn fail_with_code(code: u32, msg: String) -> Self {
         Self {
             code,
@@ -68,13 +65,11 @@ impl<T> ApiResult<T> {
     }
 
     /// Determine if ApiResult indicates successful return
-    #[inline]
     pub fn is_ok(&self) -> bool {
         self.code == 200
     }
 
     /// Determine if ApiResult indicates a return failure
-    #[inline]
     pub fn is_fail(&self) -> bool {
         self.code != 200
     }
@@ -87,7 +82,6 @@ impl<T> ApiResult<T> {
                 Some(msg) => panic!("ApiResult failed: code = {}, message = {}", self.code, msg),
                 None => panic!("ApiResult failed: code = {}", self.code),
             }
-
         }
     }
 
@@ -144,12 +138,7 @@ impl Resp {
     /// ))?;
     /// ````
     pub fn resp<T: Into<Bytes>>(status: hyper::StatusCode, body: T) -> HttpResponse {
-        Ok(
-            hyper::Response::builder()
-                .status(status)
-                .header(CONTENT_TYPE, APPLICATION_JSON)
-                .body(Full::new(body.into()))?
-        )
+        make_json_resp(status, body.into())
     }
 
     /// Create a reply with ApiResult
@@ -170,13 +159,7 @@ impl Resp {
         #[cfg(feature = "english")]
         let body = serde_json::to_vec(&ar.data).context("json serialization failed")?;
 
-        Ok(
-            hyper::Response::builder()
-                .status(status)
-                .header(CONTENT_TYPE, APPLICATION_JSON)
-                .body(Full::from(body))?
-        )
-
+        make_json_resp(status, Bytes::from(body))
     }
 
     /// Create a reply message with 200
@@ -201,16 +184,12 @@ impl Resp {
     /// ))?;
     /// ````
     pub fn resp_ok<T: Into<Bytes>>(body: T) -> HttpResponse {
-        Ok(
-            hyper::Response::builder()
-                .header(CONTENT_TYPE, APPLICATION_JSON)
-                .body(Full::from(body.into()))?
-        )
+        make_json_resp(hyper::StatusCode::OK, body.into())
     }
 
     /// Create a reply message with 200, response body is empty
-    pub fn ok_with_empty() -> HttpResponse {
-        Self::resp_ok(hyper::body::Bytes::from(r#"{"code":200}"#))
+    pub fn ok_empty() -> HttpResponse {
+        make_json_resp(hyper::StatusCode::OK, Bytes::from_static(br#"{"code":200}"#))
     }
 
     /// Create a reply message with 200
@@ -232,17 +211,15 @@ impl Resp {
     ///         },
     /// }))?;
     /// ````
-    #[inline]
-    pub fn ok<T: ?Sized + Serialize>(data: &T) -> HttpResponse {
-        // Self::ok_opt(Some(data))
-        let mut w = Vec::with_capacity(256);
-        w.extend_from_slice(br#"{"code":200,"data":"#);
+    pub fn ok<T: Serialize>(data: &T) -> HttpResponse {
+        let mut vec = Vec::with_capacity(256);
+        vec.extend_from_slice(br#"{"code":200,"data":"#);
         #[cfg(not(feature = "english"))]
-        serde_json::to_writer(&mut w, data).context("json序列化失败")?;
+        serde_json::to_writer(&mut vec, data).context("json序列化失败")?;
         #[cfg(feature = "english")]
-        serde_json::to_writer(&mut w, data).context("json serialization failed")?;
-        w.push(b'}');
-        Self::resp_ok(w)
+        serde_json::to_writer(&mut vec, data).context("json serialization failed")?;
+        vec.push(b'}');
+        make_json_resp(hyper::StatusCode::OK, Bytes::from(vec))
     }
 
     /// Create a reply message with 200
@@ -251,11 +228,10 @@ impl Resp {
     ///
     /// * `data`: http response for ApiResult.data
     ///
-    #[inline]
-    pub fn ok_opt<T: ?Sized + Serialize>(data: Option<&T>) -> HttpResponse {
+    pub fn ok_opt<T: Serialize>(data: Option<&T>) -> HttpResponse {
         match data {
             Some(v) => Self::ok(v),
-            None => Self::ok_with_empty(),
+            None => Self::ok_empty(),
         }
     }
 
@@ -272,7 +248,6 @@ impl Resp {
     ///
     /// Resp::fail("required field `username`")?;
     /// ````
-    #[inline]
     pub fn fail(message: &str) -> HttpResponse {
         Self::fail_with_code(500, message)
     }
@@ -291,7 +266,6 @@ impl Resp {
     ///
     /// Resp::fail_with_code(10086, "required field `username`")?;
     /// ````
-    #[inline]
     pub fn fail_with_code(code: u32, message: &str) -> HttpResponse {
         Self::fail_with_status(hyper::StatusCode::INTERNAL_SERVER_ERROR, code, message)
     }
@@ -315,16 +289,16 @@ impl Resp {
     pub fn fail_with_status(status: hyper::StatusCode, code: u32, message: &str) -> HttpResponse {
         let mut buf = itoa::Buffer::new();
         let code = buf.format(code);
-        let mut w = Vec::with_capacity(256);
-        w.extend_from_slice(br#"{"code":"#);
-        w.extend_from_slice(code.as_bytes());
-        w.extend_from_slice(br#","message":"#);
+        let mut vec = Vec::with_capacity(256);
+        vec.extend_from_slice(br#"{"code":"#);
+        vec.extend_from_slice(code.as_bytes());
+        vec.extend_from_slice(br#","message":"#);
         #[cfg(not(feature = "english"))]
-        serde_json::to_writer(&mut w, message).context("json序列化失败")?;
+        serde_json::to_writer(&mut vec, message).context("json序列化失败")?;
         #[cfg(feature = "english")]
-        serde_json::to_writer(&mut w, message).context("json serialization failed")?;
-        w.push(b'}');
-        Self::resp(status, w)
+        serde_json::to_writer(&mut vec, message).context("json serialization failed")?;
+        vec.push(b'}');
+        make_json_resp(status, Bytes::from(vec))
     }
 
     /// Create a reply message with specified http status and error code
@@ -335,8 +309,41 @@ impl Resp {
     /// * `code`: http error code
     /// * `message`: http error message
     ///
-    #[inline]
     pub fn internal_server_error() -> HttpResponse {
         Self::fail("internal server error")
     }
+
+    #[cfg(feature = "sse")]
+    pub fn create_sse() -> anyhow::Result<(Response, UnboundedSender<Bytes>)> {
+        use tokio_stream::{wrappers::UnboundedReceiverStream, StreamExt};
+
+        let (tx, rx) = unbounded_channel::<Bytes>();
+        let rx_stream = UnboundedReceiverStream::new(rx);
+
+        // 将Bytes转换为Frame
+        let body_stream = rx_stream.map(|v| Ok(Frame::data(v)));
+        // 将数据流转换为 StreamBody
+        let body = BoxBody::new(StreamBody::new(body_stream));
+
+        // 构造响应
+        let response = hyper::Response::builder()
+            .header("Content-Type", "text/event-stream")
+            .header("Cache-Control", "no-cache")
+            .header("Connection", "keep-alive")
+            .body(body)?;
+
+        Ok((response, tx))
+    }
+}
+
+pub(crate) fn bytes_to_body(bytes: Bytes) -> BoxBody<Bytes, Error> {
+    Full::new(bytes).map_err(map_infallible).boxed()
+}
+
+fn make_json_resp(status: hyper::StatusCode, bytes: Bytes) -> HttpResponse {
+    let res = hyper::Response::builder()
+        .status(status)
+        .header(CONTENT_TYPE, APPLICATION_JSON)
+        .body(bytes_to_body(bytes))?;
+    Ok(res)
 }
