@@ -2,57 +2,80 @@
 use std::{sync::Arc, time::Duration};
 
 use crate::{
-    auth, dict, proxy::{self, ServiceGroup}, ratelimit::RateLimiterMiddleware,
-    redis, staticmut::StaticMut, AppConf, AppGlobal
+    appvars, db, dict, efmt, proxy::{self, ServiceGroup},
+    ratelimit::RateLimiter, statics::StaticVal, AppConf
 };
-use httpserver::{http_bail, http_error, log_error, log_info, HttpContext, HttpResponse, Resp};
-use localtime::LocalTime;
+use anyhow::Context;
+use base64::{engine::general_purpose, Engine};
+use compact_str::{CompactString, format_compact};
+use httpserver::{HttpContext, HttpResponse, HttpServer, Resp, http_bail};
+use localtime::LocalDateTime;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-pub static mut RATE_LIMITER: StaticMut<Arc<RateLimiterMiddleware>> = StaticMut::new();
+pub static RATE_LIMITER: StaticVal<Arc<RateLimiter>> = StaticVal::new();
 
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 struct RegRequest {
-    endpoint: String,
-    path: Option<String>,
-    paths: Option<Vec<String>>,
+    endpoint: CompactString,
+    path: Option<CompactString>,
+    paths: Option<Vec<CompactString>>,
+}
+
+pub fn register_apis(srv: &mut HttpServer, gw_path: &str) {
+    httpserver::register_apis!(srv, gw_path,
+        "ping": ping,
+        "ping/*": ping,
+        "token": token,
+        "blacklist": blacklist,
+        "status": status,
+        "query": query,
+        "query/*": query,
+        "reg": reg,
+        "unreg": unreg,
+        "cfg": cfg,
+        "cfg/*": cfg,
+        "recfg": recfg,
+        "rate": rate,
+        "rates": rates,
+    );
 }
 
 /// 服务测试，测试服务是否存活
-pub async fn ping(ctx: HttpContext) -> HttpResponse {
+pub async fn ping(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Serialize)]
     #[serde(rename_all = "camelCase")]
     struct Res {
-        reply: String,
-        now: LocalTime,
-        server: String,
-        ip: String,
+        reply: CompactString,
+        now: LocalDateTime,
+        server: CompactString,
+        ip: CompactString,
     }
 
-    let ip = format!("{}", ctx.addr);
-    log_info!(ctx.id, "ping from: {}", ip);
+    let ip = format_compact!("{}", ctx.addr);
+    log::info!("ping from: {}", ip);
 
-    let reply = ctx.get_param_from_multi("reply", Some(0))
-        .map(String::from)
-        .unwrap_or(String::from("pong"));
+    let reply = match ctx.get_param_from_multi("reply", Some(0)).await {
+        Ok(Some(v)) => v,
+        _ => CompactString::new("pong"),
+    };
 
     Resp::ok(&Res {
         reply,
-        now: LocalTime::now(),
-        server: format!("{}/{}", crate::APP_NAME, crate::APP_VER),
+        now: LocalDateTime::now(),
+        server: format_compact!("{}/{}", crate::APP_NAME, crate::APP_VER),
         ip,
     })
 }
 
 /// 生成token，生成jwt格式token
-pub async fn token(ctx: HttpContext) -> HttpResponse {
+pub async fn token(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct Req {
-        ttl:    u32,        // token存活时间(分钟为单位)
-        claims: Value,      // 附加要加入jwt的字段及内容
+        ttl: u32,      // token存活时间(分钟为单位)
+        claims: Value, // 附加要加入jwt的字段及内容
     }
 
     #[derive(Serialize)]
@@ -66,20 +89,21 @@ pub async fn token(ctx: HttpContext) -> HttpResponse {
         return Resp::fail("jwt token generation is not supported");
     }
 
-    let param: Req = ctx.parse_json()?;
+    let param: Req = ctx.parse_json().await?;
     if !param.claims.is_object() {
         return Resp::fail("jwt token claims must be object");
     }
 
-    log_info!(ctx.id, "create jwt token from {}", param.claims);
+    log::info!("create jwt token from {}", param.claims);
     let exp = (param.ttl * 60) as u64;
-    let token = jwt::encode(param.claims, &ac.jwt_key, &ac.jwt_issuer, exp)?;
+    let token = jwt::encode(param.claims, &ac.jwt_key, &ac.jwt_issuer, exp)
+        .with_context(|| efmt!("jwt encode error"))?;
 
     Resp::ok(&Res { token })
 }
 
 /// 将token加入黑名单
-pub async fn blacklist(ctx: HttpContext) -> HttpResponse {
+pub async fn blacklist(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct Req {
@@ -87,26 +111,21 @@ pub async fn blacklist(ctx: HttpContext) -> HttpResponse {
     }
 
     let ac = AppConf::get();
-    if ac.redis_host.is_empty() || ac.jwt_key.is_empty() {
+    if ac.jwt_key.is_empty() {
         http_bail!("Not Support Jwt Token");
     }
+    if !db::is_valid() {
+        http_bail!("Not support check invalid token");
+    }
 
-    let param: Req = ctx.parse_json()?;
-    log_info!(ctx.id, "set token {} to blacklist", param.token);
+    let param: Req = ctx.parse_json().await?;
+    log::info!("set token {} to blacklist", param.token);
 
-    if let Err(e) = jwt::decode(&param.token, &ac.jwt_key, &ac.jwt_issuer) {
-        log_error!(ctx.id, "decode token fail: {e:?}");
-        http_bail!("Invalid Token");
-    };
+    if let Some((sign, exp)) = parse_token_sign_exp(&param.token) {
+        db::put(sign, exp)?;
+    }
 
-    let now = localtime::LocalTime::now();
-    let sign = jwt::get_sign(&param.token).ok_or_else(|| http_error!("Invalid Token"))?;
-    let key = format!("{}:{}:{}", ac.redis_prefix, auth::LOGOUT_KEY, sign);
-    let ttl = AppGlobal::get().redis_ttl as u32;
-
-    redis::set(&key, &now.to_string(), ttl).await?;
-
-    Resp::ok_with_empty()
+    Resp::ok_empty()
 }
 
 /// 服务状态
@@ -114,22 +133,22 @@ pub async fn status(_ctx: HttpContext) -> HttpResponse {
     #[derive(Serialize)]
     #[serde(rename_all = "camelCase")]
     struct Res {
-        startup: LocalTime,          // 应用启动时间
+        startup: LocalDateTime,      // 应用启动时间
         service_ttl: u64,            // 服务过期时间（单位：秒）
         services: Vec<ServiceGroup>, // 有效服务列表
     }
 
-    let app_global = AppGlobal::get();
+    let app_global = appvars::get();
 
     Resp::ok(&Res {
-        startup: LocalTime::from_unix_timestamp(app_global.startup_time as i64),
+        startup: LocalDateTime::from_unix_timestamp(app_global.startup_time as i64),
         service_ttl: app_global.heart_break_live_time as u64,
         services: proxy::service_status(),
     })
 }
 
 /// 注册服务查询
-pub async fn query(ctx: HttpContext) -> HttpResponse {
+pub async fn query(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Deserialize, Default)]
     #[serde(rename_all = "camelCase")]
     struct Req {
@@ -141,7 +160,7 @@ pub async fn query(ctx: HttpContext) -> HttpResponse {
     #[serde(rename_all = "camelCase")]
     struct SvrInfoItem {
         path: String,
-        services: Vec<proxy::ServiceItem>,
+        services: Vec<proxy::GroupItem>,
     }
 
     #[derive(Serialize)]
@@ -150,12 +169,12 @@ pub async fn query(ctx: HttpContext) -> HttpResponse {
         #[serde(skip_serializing_if = "Option::is_none")]
         path: Option<String>,
         #[serde(skip_serializing_if = "Option::is_none")]
-        services: Option<Vec<proxy::ServiceItem>>,
+        services: Option<Vec<proxy::GroupItem>>,
         #[serde(skip_serializing_if = "Option::is_none")]
         list: Option<Vec<SvrInfoItem>>,
     }
 
-    let mut param = ctx.parse_json_opt::<Req>()?.unwrap_or(Req::default());
+    let mut param = ctx.parse_json_opt::<Req>().await?.unwrap_or_default();
 
     // body中没有，尝试从路径中读取
     if param.path.is_none() {
@@ -170,10 +189,10 @@ pub async fn query(ctx: HttpContext) -> HttpResponse {
 
     // 优先使用path参数
     if let Some(path) = &param.path {
-        log_info!(ctx.id, "查找 {path} 对应的服务");
+        log::info!("查找 {path} 对应的服务");
         let services = proxy::service_query(path);
         return Resp::ok(&Res {
-            path: param.path.clone(),
+            path: param.path,
             services,
             list: None,
         });
@@ -194,14 +213,14 @@ pub async fn query(ctx: HttpContext) -> HttpResponse {
         });
     }
 
-    Resp::ok_with_empty()
+    Resp::ok_empty()
 }
 
 /// 注册服务(同时也作为心跳服务使用)
-pub async fn reg(ctx: HttpContext) -> HttpResponse {
+pub async fn reg(mut ctx: HttpContext) -> HttpResponse {
     type Req = RegRequest;
 
-    let param = ctx.parse_json::<Req>()?;
+    let param = ctx.parse_json::<Req>().await?;
     if param.path.is_none() && param.paths.is_none() {
         return Resp::fail("param path and paths not find");
     }
@@ -209,8 +228,7 @@ pub async fn reg(ctx: HttpContext) -> HttpResponse {
     // 优先使用path参数进行注册
     if let Some(path) = &param.path {
         if proxy::register_service(path, &param.endpoint) {
-            log_info!(
-                ctx.id,
+            log::info!(
                 "service[{} => {}] registration successful",
                 param.endpoint,
                 path,
@@ -222,24 +240,23 @@ pub async fn reg(ctx: HttpContext) -> HttpResponse {
     if let Some(paths) = &param.paths {
         for path in paths {
             if proxy::register_service(path, &param.endpoint) {
-                log_info!(
-                    ctx.id,
+                log::info!(
                     "service[{} => {}] registration successful",
                     param.endpoint,
-                    path,
+                    path
                 );
             }
         }
     }
 
-    Resp::ok_with_empty()
+    Resp::ok_empty()
 }
 
 /// 取消服务注册
-pub async fn unreg(ctx: HttpContext) -> HttpResponse {
+pub async fn unreg(mut ctx: HttpContext) -> HttpResponse {
     type Req = RegRequest;
 
-    let param = ctx.parse_json::<Req>()?;
+    let param = ctx.parse_json::<Req>().await?;
 
     if param.path.is_none() && param.paths.is_none() {
         return Resp::fail("param path and paths not find");
@@ -249,21 +266,21 @@ pub async fn unreg(ctx: HttpContext) -> HttpResponse {
 
     if let Some(path) = &param.path {
         proxy::unregister_service(path, endpoint);
-        log_info!(ctx.id, "unregister service[{}: {}]", path, endpoint);
+        log::info!("unregister service[{}: {}]", path, endpoint);
     }
 
     if let Some(paths) = &param.paths {
         for path in paths {
             proxy::unregister_service(path, endpoint);
-            log_info!(ctx.id, "unregister server[{}: {}]", path, endpoint);
+            log::info!("unregister server[{}: {}]", path, endpoint);
         }
     }
 
-    Resp::ok_with_empty()
+    Resp::ok_empty()
 }
 
 /// 获取配置信息
-pub async fn cfg(ctx: HttpContext) -> HttpResponse {
+pub async fn cfg(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct Req {
@@ -278,9 +295,12 @@ pub async fn cfg(ctx: HttpContext) -> HttpResponse {
         config: Option<dict::DictItems>,
     }
 
-    let mut param = match ctx.parse_json_opt::<Req>()? {
+    let mut param = match ctx.parse_json_opt::<Req>().await? {
         Some(v) => v,
-        None => Req { key: None, keys: None },
+        None => Req {
+            key: None,
+            keys: None,
+        },
     };
 
     if param.key.is_none() {
@@ -309,21 +329,21 @@ pub async fn cfg(ctx: HttpContext) -> HttpResponse {
 }
 
 /// 重新加载配置信息
-pub async fn recfg(ctx: HttpContext) -> HttpResponse {
+pub async fn recfg(_ctx: HttpContext) -> HttpResponse {
     let ac = AppConf::get();
     if !ac.dict_file.is_empty() {
         dict::load(&ac.dict_file).unwrap();
-        log_info!(ctx.id, "dict-file reload completed");
-        Resp::ok_with_empty()
+        log::info!("dict-file reload completed");
+        Resp::ok_empty()
     } else {
-        log_error!(ctx.id, "reload dict-file error: arg dict-file is no specified");
+        log::error!("reload dict-file error: arg dict-file is no specified");
         Resp::fail("arg dict-file no specified")
     }
 }
 
 /// 设置接口限流，rate为毫秒为单位的令牌产生速率，例如1000表示每秒产生1个令牌
 /// rate为0时，表示删除该限流器
-pub async fn rate(ctx: HttpContext) -> HttpResponse {
+pub async fn rate(mut ctx: HttpContext) -> HttpResponse {
     #[derive(Deserialize)]
     #[serde(rename_all = "camelCase")]
     struct Req {
@@ -331,16 +351,16 @@ pub async fn rate(ctx: HttpContext) -> HttpResponse {
         rate: u64,
     }
 
-    let param = ctx.parse_json::<Req>()?;
+    let param = ctx.parse_json::<Req>().await?;
 
-    let rate_limiter = unsafe { RATE_LIMITER.get() };
+    let rate_limiter = RATE_LIMITER.get();
     if param.rate == 0 {
         rate_limiter.delete(&param.path);
     } else {
         rate_limiter.insert(param.path, Duration::from_millis(param.rate));
     }
 
-    Resp::ok_with_empty()
+    Resp::ok_empty()
 }
 
 /// 查询所有限流器
@@ -353,13 +373,45 @@ pub async fn rates(_ctx: HttpContext) -> HttpResponse {
     }
     type Res = Vec<ResItem>;
 
-    let map = unsafe { RATE_LIMITER.get().get() };
-    let res: Res = map.iter()
+    let map = RATE_LIMITER.get().get();
+    let res: Res = map
+        .iter()
         .map(|item| ResItem {
             path: item.key().clone(),
-            rate: item.value().rate.as_millis() as u64,
+            rate: item.value().rate_limit.as_millis() as u64,
         })
         .collect();
 
     Resp::ok(&res)
+}
+
+fn parse_token_sign_exp(token: &str) -> Option<(&str, u64)> {
+    fn parse_exp(b64: &str) -> Option<u64> {
+        let bytes = general_purpose::URL_SAFE_NO_PAD.decode(b64.as_bytes());
+        if let Ok(bytes) = bytes {
+            if let Ok(claims) = serde_json::from_slice::<Value>(&bytes) {
+                if let Some(exp) = claims.get("exp") {
+                    if let Some(exp) = exp.as_u64() {
+                        return Some(exp);
+                    }
+                }
+            }
+        }
+        None
+    }
+
+    let mut iter = token.split('.');
+    // 跳过头部
+    if iter.next().is_some() {
+        // 读取claims部分
+        if let Some(claims_b64) = iter.next() {
+            if let Some(exp) = parse_exp(claims_b64) {
+                // 读取签名部分
+                if let Some(sign) = iter.next() {
+                    return Some((sign, exp));
+                }
+            }
+        }
+    }
+    None
 }
